// src/lib/api/generated/models/sake.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Sake Hack Backend API
 * Sake Hack Backend API

 * OpenAPI spec version: 1.0.0
 */
import type { SakeCategory } from "./sakeCategory";

export type Sake = {
  /** 酒ID */
  id: number;
  category: SakeCategory;
  /** 酒名 */
  name: string;
  /** 画像URL formatは実装次第 */
  imagePreview: string;
  /** いいね数 */
  likeCount: number;
  /** リクエストユーザーがいいね済みかどうか */
  isLiked: boolean;
};